ApartmentMS_Backend/requirements.txt
Flask==3.0.3
Flask_Cors==4.0.1
flask_sqlalchemy==3.1.1
psycopg2==2.9.9
psycopg2_binary==2.9.9
Werkzeug==3.0.3
python-dotenv